+++ raiden_contracts/data/source/raiden/Utils.sol
@@ -4,8 +4,6 @@ pragma solidity 0.5.4;
 /// @notice Utils contract for various helpers used by the Raiden Network smart
 /// contracts.
 contract Utils {
-    string constant public contract_version = "0.16.0";
-
     enum MessageTypeId {
         None,
         BalanceProof,